add 'category_id' column to events.
Add a breaking change to the database schema that adds a new column to
event. 'Categories' are like 'types,' but broader. So while types might
be things like "player talked to an NPC" or "user clicked the save
button" or "application encountered an error," categories are things
like "user events" and "application events."

The reference client was updated, but not the javascript client.
The protocol version number changed to represent this breaking change.